+++ src/components/Dropdown/DropdownItem/DropdownItem.js
@@ -1,8 +1,8 @@
 import React, { useRef, useEffect, memo } from 'react'
 import propTypes from 'prop-types'
-import { Tooltip, Checkbox } from '../../../index'
 import classNames from 'classnames'
-import useIsOverflowing from '../../../hooks/UseIsOverflowing/UseIsOverflowing'
+import { Tooltip, Checkbox } from '../../../index'
+import { useIsOverflowing } from '../../../hooks'
 import styles from './DropdownItem.module.scss'
 
 const DropdownItem = ({
@@ -31,7 +31,7 @@ const DropdownItem = ({
         item.offsetTop -
         menu.offsetTop +
         item.clientHeight -
-        (2 / 3) * menu.offsetHeight
+        (1 / 3) * menu.offsetHeight
     }
   }, [labelRef, menuRef, isFocusedByKeyboard])
 
@@ -47,7 +47,7 @@ const DropdownItem = ({
         )}
         <div ref={labelRef} className={styles.content}>
           {valueRender
-            ? valueRender(option[displayValue])
+            ? valueRender(option[displayValue], option)
             : option[displayValue]}
         </div>
       </li>
@@ -58,13 +58,13 @@ const DropdownItem = ({
 DropdownItem.propTypes = {
   /** Dropdown option. */
   option: propTypes.object,
-  /** Value to display from option. Default is 'value'. */
+  /** Value to display from option. */
   displayValue: propTypes.string,
   /** Determines if the dropdown item is focused by user's keyboard. */
   isFocusedByKeyboard: propTypes.bool,
   /** Set if multi selection is enabled. */
   multiple: propTypes.bool,
-  /** Called when selected value has clicked. */
+  /** A callback triggered whenever the user clicked on the item. */
   onClick: propTypes.func,
   /** Whether the option is selected or not. */
   isSelected: propTypes.bool,